requirements.txt
hypha_rpc==0.20.43
hypha==0.20.42
numpy==2.1.3
pillow==11.0.0
faiss-cpu==1.9.0.post1
torch==2.5.1
segment-anything==1.0
kaibu_utils==0.1.14
requests==2.32.3
python-dotenv==1.0.1
opencv-python==4.10.0.84
fastapi==0.106.0
pyvips
clip @ git+https://github.com/openai/CLIP.git